fix: lint and typing clean-up; make CI lint runnable
- mypy: Pillow/pypdf annotations (no behaviour change)
- pdf.form_fill: drop unsupported append=True (TypeError on pypdf>=6.15)
- bandit: config key is 'skips'; also skip B404
- CI: uv sync --group dev, skip gitleaks hook (no binary / nothing staged)

+++ src/media_tools/tools/video.py
@@ -510,7 +510,7 @@ def _rframe(rate: str) -> str:
         return rate
 
 
-def _human_size(size: int) -> str:
+def _human_size(size: float) -> str:
     """Format bytes to human-readable string."""
     for unit in ("B", "KB", "MB", "GB", "TB"):
         if size < 1024:
